adding task 2

+++ 2-selection_sort.c
@@ -6,8 +6,8 @@
  */
 void selection_sort(int *array, size_t size)
 {
-	size_t i, j;
-	int min, tmp;
+	size_t  i, j;
+	size_t min, tmp;
 
 	if (array == NULL)
 		return;
